Show recipient count when sending message
