+++ src/domain/messages/detail/MessagesDetail.tsx
@@ -18,6 +18,7 @@ import { Message_message as Message } from '../../../api/generatedTypes/Message'
 import { toDateTimeString, toShortDateTimeString } from '../../../common/utils';
 import KukkuuDetailPage from '../../../common/components/kukkuuDetailPage/KukkuuDetailPage';
 import useLanguageTabs from '../hooks/useLanguageTabs';
+import MessageRecipientCountField from '../fields/MessageRecipientCountField';
 import { recipientSelectionChoices } from '../choices';
 import MessageSendButton from './MessageSendButton';
 
@@ -85,37 +86,6 @@ const MessageDetailToolbar = ({
   );
 };
 
-const useRecipientCountFieldStyles = makeStyles((theme) => ({
-  container: {
-    fontSize: theme.typography.fontSize,
-    fontFamily: theme.typography.fontFamily,
-  },
-  value: {
-    fontWeight: theme.typography.fontWeightBold,
-  },
-}));
-
-const RecipientCountField = ({ record }: any) => {
-  const t = useTranslate();
-  const classes = useRecipientCountFieldStyles();
-
-  const sentAt = record?.sentAt;
-  const recipientCount = record?.recipientCount || '?';
-
-  // If message is sent, the recipient count is shown under the message
-  // title and we don't need to show it in the message body.
-  if (sentAt) {
-    return null;
-  }
-
-  return (
-    <div className={classes.container}>
-      <span className={classes.value}>{recipientCount}</span>{' '}
-      {t('messages.fields.recipientCount.label2')}
-    </div>
-  );
-};
-
 const useStyles = makeStyles((theme) => ({
   showLayout: {
     display: 'grid',
@@ -159,7 +129,7 @@ const MessagesDetail = (props: ResourceComponentPropsWithId) => {
     >
       <SimpleShowLayout className={classes.showLayout}>
         {languageTabsComponent}
-        <RecipientCountField />
+        <MessageRecipientCountField />
         <SelectField
           source="recipientSelection"
           label="messages.fields.recipientSelection.label"
